requirements.txt
pillow~=11.0.0
pypdf~=5.0.1
